Fixing and optimizing. - Error handling middlewares fixed to respond proper messages on mutiple errors catching situations (bulkCreate, bulkUpdate). - Fixed: Record delete endpoint wasn't using userID from token. - Fixed: Logic / potentially conflictive situations: User relation is now being validated for records and funds services.

+++ src/routes/recordRouter.js
@@ -56,7 +56,8 @@ router.delete('/:id',
   async (req, res, next) => {
     try {
       const { id } = req.params;
-      const data = await service.delete({ id });
+      const { sub: userID } = req.user;
+      const data = await service.delete({ id, userID });
       res.json(data);
     } catch (error) {
       next(error);
